chore: simplify helper funcs file and refactor form clean up helper in route actions
